LayoutTests/imported/w3c/web-platform-tests/navigation-api/commit-behavior/after-transition-traverse-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'navigation.currentEntry.index')

Harness Error (FAIL), message = TypeError: null is not an object (evaluating 'navigation.currentEntry.index')
